producer/requirements.txt
kafka-python
faker
requests
PyJWT
opentelemetry-api
opentelemetry-sdk
opentelemetry-instrumentation-kafka-python
opentelemetry-exporter-jaeger
